organize validators
