+++ addon/validations/nypr-accounts/custom-messages.js
@@ -2,7 +2,7 @@ export default {
   passwordRequired:      'please enter a password',
   emailRequired:         'please enter your email',
   confirmEmailRequired:  'please enter a confirmation email',
-  firstNameRequired:     'please enter your first name!',
+  firstNameRequired:     'please enter your first name',
   firstNameMaxLength:    'your first name must be {max} characters or less',
   lastNameRequired:      'please enter your last name',
   lastNameMaxLength:     'your last name must be {max} characters or less',
@@ -14,4 +14,6 @@ export default {
   userDisabled:          'This email is associated with a disabled account. If you would like to re-enable it, please contact listener services at <a href="http://www.wnyc.org/feedback" target="_blank">http://www.wnyc.org/feedback</a>.',
   loginError:            "There's an error logging in.</br><a href='/forgot'>Forget your password</a>? Need to <a href='/signup'>create an account</a>?",
   emailNotFound:         email => `we can't find an account for the email ${email}. <a href="/signup">Sign up?</a>`,
+  publicHandleRequired:  'public handle cannot be blank',
+  publicHandleExists:    'public handle already exists',
 };
