+++ app/validations/nypr-accounts/basic-card.js
@@ -1,0 +1,1 @@
+export { default } from 'nypr-account-settings/validations/nypr-accounts/basic-card';
