+++ app/validations/nypr-accounts/custom-messages.js
@@ -1,0 +1,1 @@
+export { default } from 'nypr-account-settings/validations/nypr-accounts/custom-messages';
