+++ app/validations/nypr-accounts/password.js
@@ -1,0 +1,1 @@
+export { default } from 'nypr-account-settings/validations/nypr-accounts/password';
