+++ app/validations/nypr-accounts/signup.js
@@ -1,0 +1,1 @@
+export { default } from 'nypr-account-settings/validations/nypr-accounts/signup';
