--- app/validators/nypr-accounts/basic-card.js
@@ -1,1 +1,0 @@
-export { default } from 'nypr-account-settings/validators/nypr-accounts/basic-card';
