--- app/validators/nypr-accounts/password.js
@@ -1,1 +1,0 @@
-export { default } from 'nypr-account-settings/validators/password';
