+++ app/validators/nypr-accounts/remote.js
@@ -1,1 +1,1 @@
-export { default } from 'nypr-account-settings/validators/remote';
+export { default } from 'nypr-account-settings/validators/nypr-accounts/remote';
